Add gif button which displays gif picker.

+++ lib/widgets/chat_screen/chat_screen.dart
@@ -1,6 +1,7 @@
 import 'package:adaptive_library/adaptive_library.dart';
 import 'package:dash_chat/dash_chat.dart';
 import 'package:flutter/material.dart';
+import 'package:giphy_picker/giphy_picker.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:provider/provider.dart';
 
@@ -66,6 +67,22 @@ class _ChatScreenState extends State<ChatScreen> {
                 ),
                 onSend: (message) => chatService.sendMessage(message.toJson(), messageId: message.id),
                 trailing: <Widget>[
+                  IconButton(
+                    icon: Icon(Icons.gif),
+                    onPressed: () async {
+                      final gif = await GiphyPicker.pickGif(
+                        context: context,
+                        apiKey: 'appSdG9sasG71xDf0NqEEufRt0epuE57',
+                        showPreviewPage: false,
+                      );
+
+                      if (gif != null) {
+                        final gifUrl = gif.images.original.url;
+                        final message = ChatMessage(text: '', user: user, image: gifUrl);
+                        chatService.sendMessage(message.toJson(), messageId: message.id);
+                      }
+                    },
+                  ),
                   IconButton(
                     icon: Icon(Icons.camera_alt),
                     onPressed: () => onSendImage(ImageSource.camera),
@@ -140,7 +157,8 @@ class _ChatScreenState extends State<ChatScreen> {
   }
 
   void deleteMessage(ChatMessage message) async {
-    if (message.image != null) {
+    // HACK to avoid trying to delete images from giphy which are not stored on firebase
+    if (message.image != null && !message.image.contains('giphy')) {
       await chatService.deleteFile(message.image);
     }
